Remove redundant User api and use django user model instead. Refactor user serializer to use first_name field instead of login.
