+++ exampleapp/dinosaurs/models.py
@@ -1,17 +1,18 @@
 from django.db import models
-
+from django.conf import settings
 
 class Dinosaur(models.Model):
     species = models.TextField()
 
 class User(models.Model):
-    login = models.TextField()
+    first_name = models.TextField()
     email = models.TextField()
     password = models.TextField()
 
 class Folder(models.Model):
     name = models.TextField()
     parent = models.IntegerField(null=True)
+    author = models.ForeignKey(User)
 
 class Note(models.Model):
     parent = models.IntegerField(null=True)
