+++ exampleapp/dinosaurs/serializers.py
@@ -1,6 +1,6 @@
-from dinosaurs.models import Dinosaur, User, Folder, Note
+from dinosaurs.models import Dinosaur, Folder, Note
 from rest_framework import serializers
-
+from django.contrib.auth.models import User, Group
 
 class DinosaurSerializer(serializers.HyperlinkedModelSerializer):
     class Meta:
@@ -10,7 +10,7 @@ class DinosaurSerializer(serializers.HyperlinkedModelSerializer):
 class UserSerializer(serializers.HyperlinkedModelSerializer):
     class Meta:
         model = User
-        fields = ('id', 'password', 'login', 'email')
+        fields = ('id', 'password', 'first_name', 'email')
 
 class FolderSerializer(serializers.HyperlinkedModelSerializer):
     class Meta:
@@ -21,3 +21,12 @@ class NoteSerializer(serializers.HyperlinkedModelSerializer):
     class Meta:
         model = Note
         fields = ('id', 'parent', 'name', 'text')
+
+# class UserSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = User
+
+
+class GroupSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Group
