+++ exampleapp/dinosaurs/views.py
@@ -1,7 +1,8 @@
 from rest_framework import viewsets
 from rest_framework.decorators import detail_route
-from dinosaurs.serializers import DinosaurSerializer, UserSerializer, FolderSerializer, NoteSerializer
-from dinosaurs.models import Dinosaur, User, Folder, Note
+from dinosaurs.serializers import DinosaurSerializer, UserSerializer, FolderSerializer, NoteSerializer, GroupSerializer
+from dinosaurs.models import Dinosaur, Folder, Note
+from django.contrib.auth.models import User, Group
 import pdb
 from django.http import HttpResponse, JsonResponse
 from rest_framework import permissions
@@ -17,10 +18,10 @@ class DinosaurViewSet(viewsets.ModelViewSet):
     queryset = Dinosaur.objects.all()
     serializer_class = DinosaurSerializer
 
-class UserViewSet(viewsets.ModelViewSet):
-    permission_classes = [permissions.AllowAny]
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
+# class UserViewSet(viewsets.ModelViewSet):
+#     permission_classes = [permissions.AllowAny]
+#     queryset = User.objects.all()
+#     serializer_class = UserSerializer
 
     # def get(self, request):
     #     pdb.set_trace()
@@ -38,9 +39,24 @@ class FolderViewSet(viewsets.ModelViewSet):
     queryset = Folder.objects.all()
 
     def list(self, request):
+        pdb.set_trace()
         serializer = FolderSerializer(Folder.objects.all(), many=True)
         return HttpResponse(json.dumps(serializer.data))
 
 class NoteViewSet(viewsets.ModelViewSet):
     queryset = Note.objects.all()
     serializer_class = NoteSerializer
+
+# # ViewSets define the view behavior.
+class UserViewSet(viewsets.ModelViewSet):
+    permission_classes = [permissions.IsAuthenticated, TokenHasReadWriteScope]
+    permission_classes = [permissions.AllowAny]
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+
+
+class GroupViewSet(viewsets.ModelViewSet):
+    permission_classes = [permissions.IsAuthenticated, TokenHasScope]
+    required_scopes = ['groups']
+    queryset = Group.objects.all()
+    serializer_class = GroupSerializer
