+++ exampleapp/exampleapp/urls.py
@@ -18,43 +18,19 @@ from rest_framework import routers
 from django.contrib import admin
 from dinosaurs import views
 
-from django.contrib.auth.models import User, Group
 admin.autodiscover()
 
 from rest_framework import permissions, routers, serializers, viewsets
 
 from oauth2_provider.contrib.rest_framework import TokenHasReadWriteScope, TokenHasScope
 
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-
-
-class GroupSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Group
-
-# # ViewSets define the view behavior.
-class UserViewSet(viewsets.ModelViewSet):
-    permission_classes = [permissions.IsAuthenticated, TokenHasReadWriteScope]
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
-
-
-class GroupViewSet(viewsets.ModelViewSet):
-    permission_classes = [permissions.IsAuthenticated, TokenHasScope]
-    required_scopes = ['groups']
-    queryset = Group.objects.all()
-    serializer_class = GroupSerializer
-
-
 router = routers.DefaultRouter()
 router.register(r'dinosaurs', views.DinosaurViewSet)
-router.register(r'user', views.UserViewSet)
+# router.register(r'user', views.UserViewSet)
 router.register(r'folder', views.FolderViewSet)
 router.register(r'note', views.NoteViewSet)
-router.register(r'users', UserViewSet)
-router.register(r'groups', GroupViewSet)
+router.register(r'users', views.UserViewSet)
+router.register(r'groups', views.GroupViewSet)
 
 # Wire up our API using automatic URL routing.
 # Additionally, we include login URLs for the browsable API.
